feat: 优化Hunyuan3D GRPO训练流程
主要改进:
- 🔧 修复多GPU分布式训练的数据同步问题
- ⚡ 优化FlashVDM配置和内存管理
- 🎯 改进采样器和批次处理逻辑
- 📊 增强训练日志和调试信息
- 🛠️ 修复扩散管道的log_prob计算
- 🚀 优化训练脚本的稳定性和性能

修改文件:
- config/hunyuan3d.py: 更新训练配置
- scripts/train_hunyuan3d.py: 核心训练逻辑优化
- generators/hunyuan3d/: 模型组件优化
- flow_grpo/: GRPO算法实现改进

+++ generators/hunyuan3d/hy3dshape/models/autoencoders/surface_extractors.py
@@ -111,7 +111,7 @@ class MCSurfaceExtractor(SurfaceExtractor):
                   box coordinates.
                 - faces (np.ndarray): Extracted mesh faces (triangles).
         """
-        vertices, faces, normals, _ = measure.marching_cubes(grid_logit.cpu().numpy(),
+        vertices, faces, normals, _ = measure.marching_cubes(grid_logit.cpu().float().numpy(),
                                                              mc_level,
                                                              method="lewiner")
         grid_size, bbox_min, bbox_size = self._compute_box_stat(bounds, octree_resolution)
